Factor out and restructure scala.meta.internal.javacp
* Create scala.meta.internal.javacp and scala.meta.internal.javacp.asm
* Move JavaTypeSignature.scala to scala.meta.internal.javacp
* Make different flavors and elements of JavaTypeSignature toplevel
* Factor out Printable into a separate file
