+++ semanticdb/metacp/src/main/scala/scala/meta/internal/javacp/Scope.scala
@@ -1,4 +1,4 @@
-package scala.meta.internal.metacp
+package scala.meta.internal.javacp
 
 /**
   * Minimal utility to resolve generic signature type variables to fully qualified symbols.
