+++ semanticdb/metacp/src/main/scala/scala/meta/internal/javacp/asm/ClassSignatureVisitor.scala
@@ -1,7 +1,7 @@
-package scala.meta.internal.metacp.asm
+package scala.meta.internal.javacp.asm
 
 import scala.collection.JavaConverters._
-import scala.meta.internal.metacp.asm.JavaTypeSignature._
+import scala.meta.internal.javacp._
 import scala.tools.asm.signature.SignatureVisitor
 
 class ClassSignatureVisitor
