+++ semanticdb/metacp/src/main/scala/scala/meta/internal/javacp/asm/JavaTypeSignatureVisitor.scala
@@ -1,7 +1,6 @@
-package scala.meta.internal.metacp.asm
+package scala.meta.internal.javacp.asm
 
-import scala.meta.internal.metacp.asm.JavaTypeSignature.BaseType
-import scala.meta.internal.metacp.asm.JavaTypeSignature.ArrayTypeSignature
+import scala.meta.internal.javacp._
 import scala.tools.asm.signature.SignatureVisitor
 
 class JavaTypeSignatureVisitor(isArray: Boolean) extends TypedSignatureVisitor[JavaTypeSignature] {
