+++ semanticdb/metacp/src/main/scala/scala/meta/internal/javacp/asm/TypedSignatureVisitor.scala
@@ -1,4 +1,4 @@
-package scala.meta.internal.metacp.asm
+package scala.meta.internal.javacp.asm
 
 /** Helper to abstract over SignatureVisitors that produce a result of type T */
 abstract class TypedSignatureVisitor[+T] extends FailFastSignatureVisitor {
